+++ src/Yasumi/data/translations/dayOfReformation.php
@@ -1,0 +1,17 @@
+<?php
+/**
+ * This file is part of the Yasumi package.
+ *
+ * Copyright (c) 2015 - 2018 AzuyaLabs
+ *
+ * For the full copyright and license information, please view the LICENSE
+ * file that was distributed with this source code.
+ *
+ * @author Sacha Telgenhof <stelgenhof@gmail.com>
+ */
+
+// Translations for Day of Reformation
+return [
+    'de_DE' => 'Tag der Reformation',
+    'en_US' => 'Day of Reformation',
+];